Implemented autopilot to position

+++ src/OpenSBS.Core/Models/EntityTrace.cs
@@ -4,14 +4,14 @@
     {
         public string Id { get; }
         public int Bearing { get; protected set; }
-        public int Distance { get; protected set; }
+        public float Distance { get; protected set; }
 
         public EntityTrace(string id)
         {
             Id = id;
         }
 
-        public void Update(int bearing, int distance)
+        public void Update(int bearing, float distance)
         {
             Bearing = bearing;
             Distance = distance;
